Basic layout for Exec Board page
Here is some basic layout for content pages. I used the Exec Board page
as an example. I was only focusing on the layout of the content, not
the header or footer. Referencing #13.

Things left to do:
* Fix column width (if I do 2 - 1/3-width columns, some browsers force
them all to the left instead of centering. If I do 2 - 1/2-width
columns, they end up larger in width than the 1/3rd-width columns)
- [ ] Move CSS and images to their own folder (priority - will rectify
tomorrow)
- [ ] Clean up HTML and CSS
- [ ] Finish header (#1)
- [ ] Finish footer (#4)
- [ ] Finish home page (also #13)

Long term, this will be used as a template and individual pages will be
made once this is completed.